chore: move character files for new packages and to use new architecture
