+++ app/src/main/java/feature/HomeActivity.kt
@@ -1,10 +1,10 @@
-package com.example.starwars.screens.home
+package feature
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.starwars.R
-import com.example.starwars.screens.character.CharactersActivity
+import feature.character.presentation.CharactersActivity
 import com.example.starwars.screens.film.FilmsActivity
 import com.example.starwars.screens.planet.PlanetsActivity
 import kotlinx.android.synthetic.main.activity_home.*
