+++ app/src/main/java/feature/character/domain/models/Character.kt
@@ -1,4 +1,4 @@
-package com.example.starwars.models
+package feature.character.domain.models
 
 import java.io.Serializable
